Add option for bidirectional encoder
Added parser argument to turn on/off the bidirectional encoder. Based on
the bool, the number of decoder hidden units is also changed (doubled if
bidirectional encoder is used). Also added parser argument to turn on or
off the use of attention over the encoder outputs. It doesn't do anything
yet though.