Update data processing
Updated the data processor, so that it is now working directly with the
raw json files. The files are read then parsed, finally they get saved
into binary files with pickle. If it is run again, it will load the
binaries directly. The make_dictionary method was also updated to work
directly with the json files when collecting vocabularies. Overall, the
startup time of scripts was greatly reduced to a few seconds now.